requirements.txt
autobahn==23.6.2
beautifulsoup4==4.12.2
discord.py==1.6.0
django-extensions==3.2.3
django-ratelimit==4.1.0
h11==0.14.0
Markdown==3.5.2
pillow==10.2.0
pip==23.3.2
psycopg2==2.9.9
pyOpenSSL==23.3.0
python-memcached==1.61
service-identity==24.1.0
twisted==23.10.0
uwuify==1.3.0
wheel==0.42.0
whitenoise==6.6.0
zope.interface==6.1